+++ lib/hostares.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: hostares.c,v 1.34 2007-10-03 13:19:34 yangtse Exp $
+ * $Id: hostares.c,v 1.35 2007-10-20 15:11:51 yangtse Exp $
  ***************************************************************************/
 
 #include "setup.h"
@@ -255,7 +255,7 @@ CURLcode Curl_wait_for_resolv(struct connectdata *conn,
     tvp = ares_timeout(data->state.areschannel, &store, &tv);
 
     /* use the timeout period ares returned to us above */
-    ares_waitperform(conn, tvp->tv_sec * 1000 + tvp->tv_usec/1000);
+    ares_waitperform(conn, (int)(tvp->tv_sec * 1000 + tvp->tv_usec/1000));
 
     if(conn->async.done)
       break;
